fixed bug with bh1750 sensor

+++ internal/sensors/devicesensors/bh1750.go
@@ -1,6 +1,6 @@
 package sensors
 
 type BH1750Measurement struct {
-	Mac       string  `json:"m"`
-	Intensity float64 `json:"i"`
+	Mac       string `json:"m"`
+	Intensity int32  `json:"i"`
 }
